note-taking-app/requirements.txt
numpy
opencv-python
opencv-contrib-python
mysql-connector-python
pyttsx3
pillow
pysimplegui
matplotlib